Moved permissions services to Authorization folder, returned back auth in program.cs
