+++ backend/Time-Tracker/Authorization/PermissionHandler.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Time_Tracker.Helpers;
-using Time_Tracker.Services;
 
 namespace Time_Tracker.Authorization
 {
