+++ backend/Time-Tracker/Time-Tracker.csproj
@@ -26,4 +26,8 @@
     <PackageReference Include="Microsoft.Data.SqlClient" Version="5.2.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Services\Permission\" />
+  </ItemGroup>
+
 </Project>
